Quick push while working

+++ commands/erase.js
@@ -1,19 +1,28 @@
 const { SlashCommandBuilder, PermissionFlagsBits } = require('discord.js');
+const { permissions } = require('../config.json');
 
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('erase')
-		.setDescription('Deletes previous messages ?')
+		.setDescription('Deletes previous messages')
 		.addStringOption(option =>
 			option
 				.setName('quantity')
-				.setDescription('How many messages to delete'))
+				.setDescription('How many messages should I delete?'))
 		.setDefaultMemberPermissions(PermissionFlagsBits.ManageMessages),
 	async execute(interaction) {
 
 		const member = interaction.member;
 
-		if (member.roles.cache.some(role => role.name === 'Moderator')) {
+		// Permission check
+		let canRun = false;
+		member.roles.cache.forEach((role) => {
+			if (permissions.erase.includes(role.name)) {
+				canRun = true;
+			}
+		});
+
+		if (canRun) {
 
 			// Acknowledge
 			await interaction.reply({ content: 'Working on it!', ephemeral: true });
@@ -23,16 +32,21 @@ module.exports = {
 
 			let count = 0;
 			channel.messages.fetch({ limit: quantity }).then(messages => {
-				// Iterate through the messages here with the variable "messages".
-				messages.forEach(message => {
-					setTimeout(() => {
-						message.delete();
-						count++;
-						if (count === messages.size) {
-							interaction.editReply(`Erased ${count} messages.`);
-						}
-					}, 1000);
-				});
+				if (messages.size > 0) {
+					// Iterate through the messages here with the variable "messages".
+					messages.forEach(message => {
+						setTimeout(() => {
+							message.delete();
+							count++;
+							if (count === messages.size) {
+								interaction.editReply(`Erasing ${count} messages, just be patient!`);
+							}
+						}, 1000);
+					});
+				}
+				else {
+					interaction.editReply('Seems like there is nothing to delete.');
+				}
 			});
 
 		}
